handles both old and new style locators.
NOTE: There might still be conditions that this fails on.
For example, if the file desired spans multiple manifest
elements, this will fail in weird ways.  Though it should
handle old-style locators with files in sub-directories (that is,
not in the root directory as is the case most of the time), this
is untested.

There should probably be an arvados tool to get the file size or
do better querying of manifests.  If and when that tool exists, this
code should be updated if it's still in use.